LPS-192496 Extract service from ConfigurationModelIndexer to ConfigurationModelIndexerHelper, and apply to usages
